REL-2859 Code refactorings and cleanup #time 30m
